% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_tz.R
\name{city_tz}
\alias{city_tz}
\title{Converts a city name to the associated timezone}
\usage{
city_tz(city)
}
\arguments{
\item{city}{Name of city. Spaces must be indicated by underscores.Not case-sensitive.}
}
\value{
Name of associated timezone
}
\description{
Converts a city name to the associated timezone
}
\examples{
city_tz("Singapore") gives "Asia/Singapore"
city_tz("Cape_Town") gives "Africa/Johannesburg"

}
