% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/City_Zone.R
\docType{data}
\name{City_Zone}
\alias{City_Zone}
\title{City timezones.}
\format{A data frame with 428 rows and 2 variables}
\usage{
data(City_Zone)
}
\description{
A dataset containing the world's major cities and their corresponding timezones. The variables are as follow:
}
\details{
\itemize{
  \item City.
  \item Zone.
}
}
\keyword{datasets}
