% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tz_conversion.R
\name{tz_conversion}
\alias{tz_conversion}
\title{Timezone_conversion converts a specified time in city1 to the time in city2.}
\usage{
tz_conversion(city2, city1 = Sys.timezone(), time = Sys.time())
}
\arguments{
\item{city2}{name of city the user wants time to be converted to}

\item{city1}{name of source time city. Default is system timezone}

\item{time}{time of source time city, format YYYY-MM-DD HH:MM:SS. Default is system time}
}
\value{
time in desired city
}
\description{
Timezone_conversion converts a specified time in city1 to the time in city2.
}
\examples{
tz_conversion("Singapore")
tz_conversion("Cape_Town", "Singapore", "2017-07-03 20:00:00")
}
